web-ui is ready without backend

+++ hlf-Dapp/chaincode-javascript/lib/dltContract.js
@@ -11,45 +11,15 @@ class DltContract extends Contract {
 
     async initLedger(ctx) {
         console.info('============= START : Initialize Ledger ===========');
-        // const assets = [
-        //     {
-        //         assetType:'project',
-        //         projectName:'Radha Krishna Apartment',
-        //         flatPrice:5000,
-        //         description:'only 3BKH flats available'       
-        //     },
-        //     {
-        //         assetType:'project',
-        //         projectName:'Sita Ram Apartment',
-        //         flatPrice:2000,
-        //         description:'only 3BKH flats available'       
-        //     },
-        //     {
-        //         assetType:'project',
-        //         projectName:'Mahavira Apartment',
-        //         flatPrice:4500,
-        //         description:'only 3BKH flats available'       
-        //     },
-        //     {
-        //         assetType:'project',
-        //         projectName:'Radhay Complex',
-        //         flatPrice:4000,
-        //         description:'only 3BKH flats available'       
-        //     },
-        // ];
-
-        // for (let i = 0; i < assets.length; i++) {
-        //     assets[i].assetType = 'project';
-        //     await ctx.stub.putState('Project' + i, Buffer.from(JSON.stringify(assets[i])));
-        //     console.info('Added <--> ', assets[i]);
-        // }
+        
         console.info('============= END : Initialize Ledger ===========');
     }
 
     
     // create project
     async createProject(ctx, args) {
-             
+        let cid = new ClientIdentity(ctx.stub);        
+        const userId = await cid.getAttributeValue('userId');    
                 
         args = JSON.parse(args);
         console.log("args: ", args);
@@ -58,35 +28,70 @@ class DltContract extends Contract {
 
         await ctx.stub.putState(args['projectId'], Buffer.from(JSON.stringify(args['project'])));
         console.info('============= END : Create Project ===========');
+
+        const transactionId = ctx.stub.getTxID();
+        // const obj = ctx.stub.getState(transactionId);
+        // console.log('transaction Object:', obj);
+        console.log('Transaction ID:', transactionId);
+
+        const eventPayload = {
+            userName:userId,
+            assetData:args['project'],
+            transactionId:transactionId          
+        }
+
+		const assetBuffer = Buffer.from(JSON.stringify(eventPayload));
+
+        ctx.stub.setEvent('CreateAssetEvent', assetBuffer);
+        console.info('============= CreateAssetEvent EVENT EMIT ===========');
+
         return {status:true, message:'Project created success'};
       
     }
 
      // update project
      async updateProject(ctx, args) {
-        let cid = new ClientIdentity(ctx.stub);        
-        const owner = await cid.getAttributeValue('userId');
-        console.info(`User Identity : ${owner}`)
-        args = JSON.parse(args);
-        console.log("args: ", args);
 
-        if(owner === args["owner"]){
-            console.info(`============= START : update Project by ${args['owner']} ===========`);
+        let cid = new ClientIdentity(ctx.stub);        
+        const userId = await cid.getAttributeValue('userId');
+        console.info(`---------------update function => User Identity : ${userId}--------`)
+        args = JSON.parse(args);
+        console.log("user : ", args.project.owner);
+
+        if(userId == args.project.owner){
+
+            console.info(`============= START : update Project by  ===========`);
 
             // owner can update asset with in 5mins // 5*60,000 = 300000 in millisecond
             const result = await ctx.stub.getState(args['projectId']);
             const oldData = JSON.parse(result.toString());
             const currentTime = Date.now();
-            const assetTimestamp = oldData.timestamp;
+            const assetTimestamp = parseInt(oldData['timestamp']);
+          
+            console.log(`assetTimestamp :: ${assetTimestamp}`)
 
-            if(currentTime - assetTimestamp <= (300000)){
+            if((currentTime - assetTimestamp) <= 300000){
                 
                 await ctx.stub.putState(args['projectId'], Buffer.from(JSON.stringify(args['project'])));
                 console.info('============= END : update Project ===========');
+                
+                const transactionId = ctx.stub.getTxID();
+
+                const eventPayload = {
+                    userName:userId,
+                    assetData:args['project'],
+                    transactionId:transactionId                    
+                };
+
+    		    const assetBuffer = Buffer.from(JSON.stringify(eventPayload));
+
+                ctx.stub.setEvent('UpdateAssetEvent', assetBuffer);
+                console.info('============= UpdateAssetEvent EVENT EMIT ===========');
+                
                 return {status:true, message:'Project updated success'};
 
             }else {
-                return {status:false, message:'owner can update project with in 5mins after creation.'};
+                    return {status:false, message:'owner can update project with in 5mins after creation.'};
                 }
 
     }else {
@@ -112,51 +117,60 @@ class DltContract extends Contract {
     // deleteProject
     async  deleteProject(ctx, args)
     {
+        let cid = new ClientIdentity(ctx.stub);        
+        const owner = await cid.getAttributeValue('userId');
+        console.info(`-------Delete function --- User Identity : ${owner} -------------`);
+
         args = JSON.parse(args);
-         await ctx.stub.deleteState(args['assetId']); // get the asset from chaincode state
-        console.log("Project deleted");
-       
-            return {status:true, data:`Project deleted - ${args['assetId']}`};
+        if(owner == args['owner']){
+
+            await ctx.stub.deleteState(args['assetId']); // get the asset from chaincode state
+           console.log("Project deleted");
+           return {status:true, data:`Project deleted - ${args['assetId']}`};
+
+        }else{
+            return {status:false, message:'only owner can delete project!!'};
+        }       
         
     }
         
 
-    // // queryAllAssetByUser
-    // async queryAllAssetByUser(ctx) {
+    // queryAllAssetByOwner
+    async queryAllProjectByOwner(ctx) {
 
-    //     // const investmentAsset = {
-    //     //     assetType:'investment',
-    //     //     userID:userID,     
-    //     //     flatPrice:flatPrice,
-    //     //     projectId:projectId,
-    //     //     projectName:projectName  
-    //     // };
+        // const asset = {
+        //     assetType:'investment',
+        //     owner:userID,     
+        //     flatPrice:flatPrice,
+        //     projectId:projectId,
+        //     projectName:projectName  
+        // };
 
-    //     let cid = new ClientIdentity(ctx.stub);
-    //     const userID = await cid.getAttributeValue('userId');
+        let cid = new ClientIdentity(ctx.stub);
+        const userID = await cid.getAttributeValue('userId');
         
-    //     const type = 'investment'; // 'project', 'wallet', 'investment'
-    //     const allResults = [];
-    //     let queryString = {};
-    //     queryString.selector = {};
-    //     queryString.selector.assetType = type;
-    //     queryString.selector.userID = userID;
+        const type = 'project'; // 'project'
+        const allResults = [];
+        let queryString = {};
+        queryString.selector = {};
+        queryString.selector.assetType = type;
+        queryString.selector.owner = userID;
 
-    //     for await (const { key, value }
-    //         of ctx.stub.getQueryResult(JSON.stringify(queryString))) {
-    //         const strValue = Buffer.from(value).toString('utf8');
-    //         let record;
-    //         try {
-    //             record = JSON.parse(strValue);
-    //         } catch (err) {
-    //             console.log(err);
-    //             record = strValue;
-    //         }
-    //         allResults.push({ assetId: key, record: record });
-    //     }
-    //     console.info(allResults);
-    //     return {status:true, data:JSON.stringify(allResults)};
-    // }
+        for await (const { key, value }
+            of ctx.stub.getQueryResult(JSON.stringify(queryString))) {
+            const strValue = Buffer.from(value).toString('utf8');
+            let record;
+            try {
+                record = JSON.parse(strValue);
+            } catch (err) {
+                console.log(err);
+                record = strValue;
+            }
+            allResults.push({ assetId: key, record: record });
+        }
+        console.info(allResults);
+        return {status:true, data:JSON.stringify(allResults)};
+    }
 
  
     // queryAllAssets on ledger
@@ -183,15 +197,16 @@ class DltContract extends Contract {
                 return {status:false, error: `${args['assetID']} does not exist`};
             } else {
                 return {status:true, assets:JSON.stringify(allResults)};
-
             }
     }
 
     
     //get history of project
-    async getHistoryOfProject(ctx, projectId) {
+    async getHistoryOfProject(ctx, args) {
        
-        for await (const {key, value} of ctx.stub.getHistoryForKey(projectId)) {
+        args = JSON.parse(args);
+
+        for await (const {key, value} of ctx.stub.getHistoryForKey(args['projectId'])) {
             const strValue = Buffer.from(value).toString('utf8');
             let record;
             try {
